Add files via upload
Moving over to fully implemented units and using the orbital elements module to determine random positions and velocities. Added the option of running Mercury, which shows better performance for N_objects ~> 2*10**3. Removing the galactic bridge caused a 5x speed up. A run up to 27000 yrs for 100 objects found 1 escaping object at 350 AU, so even with a 'vanilla-system' objects appear to be able to escape.